fix: ✏️ Users are unable to delete their accounts while they hold ownership of any chatrooms (#198)
* feat: add method of find existence that user has chatroom's ownership anyone on repository

* feat: add method in domain service

* feat: add 409 error for user's signout blocking because user has chat room ownership

* feat: resolve todo annotation

* test: add delete test about has ownership chatroom user can't delete account

+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/apis/users/service/UserDeleteService.java
@@ -1,6 +1,7 @@
 package kr.co.pennyway.api.apis.users.service;
 
 import kr.co.pennyway.domain.domains.device.service.DeviceTokenService;
+import kr.co.pennyway.domain.domains.member.service.ChatMemberService;
 import kr.co.pennyway.domain.domains.oauth.service.OauthService;
 import kr.co.pennyway.domain.domains.spending.service.SpendingCustomCategoryService;
 import kr.co.pennyway.domain.domains.spending.service.SpendingService;
@@ -27,6 +28,8 @@ public class UserDeleteService {
     private final OauthService oauthService;
     private final DeviceTokenService deviceTokenService;
 
+    private final ChatMemberService chatMemberService;
+
     private final SpendingService spendingService;
     private final SpendingCustomCategoryService spendingCustomCategoryService;
 
@@ -36,13 +39,14 @@ public class UserDeleteService {
      * hard delete가 수행되어야 할 데이터는 삭제하지 않으며, 사용자 데이터 유지 기간이 만료될 때 DBA가 수행한다.
      *
      * @param userId
-     * @todo [2024-05-03] 채팅 기능이 추가되는 경우 채팅방장 탈퇴를 제한해야 하며, 추가로 삭제될 엔티티 삭제 로직을 추가해야 한다.
      */
     @Transactional
     public void execute(Long userId) {
         if (!userService.isExistUser(userId)) throw new UserErrorException(UserErrorCode.NOT_FOUND);
 
-        // TODO: [2024-05-03] 하나라도 채팅방의 방장으로 참여하는 경우 삭제 불가능 처리
+        if (chatMemberService.hasUserChatRoomOwnership(userId)) {
+            throw new UserErrorException(UserErrorCode.HAS_OWNERSHIP_CHAT_ROOM);
+        }
 
         oauthService.deleteOauthsByUserIdInQuery(userId);
         deviceTokenService.deleteDevicesByUserIdInQuery(userId);
